feat: auto-detect precision for crypto vs fiat currencies
- Default to 8 decimal places for crypto currencies (btc, eth, etc.)
- Default to 2 decimal places for fiat currencies (usd, sgd, etc.)
- Remove $ prefix for crypto currency display
- Change f32 to f64 for better precision with small values
- Remove .idea folder from version control

+++ src/lib.rs
@@ -1,4 +1,4 @@
 
-pub fn calculate_fees(fees: f32, current_holdings: f32) -> f32 {
+pub fn calculate_fees(fees: f64, current_holdings: f64) -> f64 {
     current_holdings * fees
 }
